+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- ACF Codifier Components
+    + Components are reusable field sets.
+    + Components can be transformed into any type of groupable field with component factories.
+        * `Geniem\ACF\Factory\Block`: A factory for creating ACF Gutenberg blocks with Codifier components.
+    + Component rendering functionality implementing the `Geniem\ACF\Interfaces\Renderer` interface.
+        * `Geniem\ACF\Renderer\PHP`: Uses PHP files as templates.
+
 ## [1.14.3] - 2018-10-18
 
 ### Fixed